most active crypto/error handling/ firebase db fix

+++ components/stockdetails/watchList.js
@@ -3,14 +3,18 @@ import { TouchableOpacity } from 'react-native';
 import Icon from 'react-native-vector-icons/MaterialIcons';
 import { doc, setDoc, deleteDoc, getDoc } from "firebase/firestore"; 
 import { db } from '../../config/firebase'; 
+import { getAuth, deleteUser } from "firebase/auth";
 
 const WatchList = ({ symbol, name, price, type, color }) => {
   const [isFavorited, setIsFavorited] = useState(false);
+  const auth = getAuth(); 
+  const user = auth.currentUser;
 
   useEffect(() => {
+    if (!user) return;
     // Check if the stock is already favorited in Firestore
     const checkFavoriteStatus = async () => {
-      const docRef = doc(db, "watchList", symbol);
+      const docRef = doc(db, `users/${user.uid}/watchList`, symbol);
       const docSnap = await getDoc(docRef);
       if (docSnap.exists()) {
         setIsFavorited(true);
@@ -18,10 +22,12 @@ const WatchList = ({ symbol, name, price, type, color }) => {
     };
 
     checkFavoriteStatus();
-  }, [symbol]);
+  }, [symbol, user]);
 
   const toggleFavorite = async () => {
-    const docRef = doc(db, "watchList", symbol);
+    if (!user) return;
+
+    const docRef = doc(db, `users/${user.uid}/watchList`, symbol);
     if (isFavorited) {
       // Remove from favorites
       await deleteDoc(docRef);
